backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-functions>=1.23.0",
    "conllu>=6.0.0",
    "dotenv>=0.9.9",
]

[tool.uv.workspace]
members = [
    "data",
]
